backend/requirements.txt
asgiref==3.5.2
cffi==1.15.1
Django==4.1.3
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
grpcio==1.50.0
grpcio-tools==1.48.2
iroha==1.4.1.1
Pillow==9.3.0
protobuf==3.20.1
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pytz==2022.6
segno==1.5.2
six==1.16.0
sqlparse==0.4.3
